+++ src/components/spartan/SSidebar/index.ts
@@ -1,5 +1,6 @@
 export { default as SSidebar } from './SSidebar.vue';
 export { default as SSidebarItem } from './SSidebarItem.vue';
 export { default as SSidebarItemGroup } from './SSidebarItemGroup.vue';
+export { default as SSidebarSeparator } from './SSidebarSeparator.vue';
 
-export type { TSidebarEmits, TSidebarProps, TSidebarItemProps, TSidebarItemGroupProps } from './types';
+export type { TSidebarEmits, TSidebarProps, TSidebarItemProps, TSidebarItemGroupProps, SSidebarSeparator } from './types';